refactor: 인스턴스 간 알림 전파(Redis) 코드를 sse 패키지에서 broadcast로 분리
로컬 SSE 연결 관리(sse)와 인스턴스 간 신호 전파(broadcast) 관심사가 한 패키지에
섞여 있어 분리. 로직 변경 없음.
