+++ backend/src/main/java/org/example/be/domain/notification/broadcast/RedisPubSubNotificationSender.java
@@ -1,4 +1,4 @@
-package org.example.be.domain.notification.sse;
+package org.example.be.domain.notification.broadcast;
 
 import org.example.be.domain.notification.dto.response.NotificationResBody;
 import org.springframework.data.redis.core.StringRedisTemplate;
